Fix line-joining bug in vim minisnip plugin
Adding foldexpr to certain filetypes caused the completion of snippets
to break. After pressing tab to complete a snippet, there would be an
unpredictable joining of lines in the snippet. This was because the
append function call in the plugin source code added lines that would
automatically get folded according to the settings of foldlevel and foldexpr.
This caused any following code to behave unpredictably. To fix this, I
altered the source code to disable folding before any calls to append()
and then restore the original fold setting after the call.