Edit minisnip to not ring bell at last placeholder
Since the placeholder selection function in the minisnip plugin uses
try-catches to determine whether there are remaining placeholders, once
you reach the final placeholder or if there are no placeholders left,
the bell will sound from an expected E486 (no matching pattern). To
avoid the bell, I store the value of 'belloff' into a variable and then
append the "error" option to 'belloff' at the beginning of the
SelectPlaceholder() function.  Then, at the end of the function, I
restore the original value of belloff.