Improve foldexpr for c and cpp
Include statements and header statements get a foldlevel of zero rather
than an opening foldlevel of one. Closing brace followed by a semicolon
at gets a foldlevel of one rather than zero. Any blank lines after
functions get a foldlevel of zero rather than being included in the
function fold. This is so that `o` or `O` on a closed fold doesn't cause
the fold to open since the new line is no longer inside of the fold.