Add ysS command to vim-surround
Unlike yss, ysS includes surrounding whitespace. The main changes to the
plugin are the addition of a global variable named surround_ws that is
set to 0 by default. The ysS mapping changes this value to 1, which
causes opfunc() to skip the manipulation of trailing whitespace. The
value is then changed back to 0 at the end of opfunc() to restore stock
behavior of the plugin.