Fix folding bug in markdown ftplugin
A blank line before a heading is desired to have a foldlevel of one less
than the heading to separate closed fold by one line. The previous
method for implementing this used the foldelevel() function to get the
foldlevel of the next line and then use that to set the foldlevel of the
current line. However, this caused unpredictable behavior when a
markdown file started with a heading in the first line. We now hardcode
each case of a heading in the next line and return the foldlevel based
on that.